requirements.txt
matplotlib.pyplot
streamlit
pandas
seaborn